pyproject.toml
[tool.poetry]
name = "hedwig"
version = "0.1.0"
description = "Hedwig, the Messenger"
authors = ["Vibhu Agarwal <vibhu4agarwal@gmail.com>"]
license = "GNU GPLv3"

[tool.poetry.dependencies]
python = "^3.9"
pymongo = "^3.11.4"
fastapi = "^0.63.0"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
pymodm = "^0.4.3"
pydantic = "^1.8.1"
python-dotenv = "^0.17.1"
dnspython = "^2.1.0"
aioredis = {version = "2.0.0a1", extras = ["hiredis"]}
asgiref = "^3.3.4"

[tool.poetry.dev-dependencies]
black = "^21.5b0"
flake8 = "^3.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
